Added some dpomdp files. Made a DPOMDP gym environment based on Andrea's dpomdp parser and POMDP environment.
The DPOMDP environment works on many of the files, but runs into two classes of issues:
The first is that sometimes the parser fails. I'm going to leave that for now
The second is that due to rounding, sometimes the transition probabilities of an environment fail to sum to 1 for some permutation. I'm working on a function to normalize the tables